option to set owner
This is for the brain only. It does nothing for permissions on discord.
This is so the bot will reply with a name if asked who the creator is.
In the future maybe something could be writen to obtain the bot owner from discord itself to set these values. For now manual works for me.